+++ tests/test_pattern.py
@@ -938,14 +938,10 @@ class TestPattern:
         circuit_1 = rand_circuit(nqubits, depth, rng, use_ccx=False)
         p_ref = circuit_1.transpile().pattern
         p_ref.infer_pauli_measurements()
-        p_test = p_ref.to_bloch().to_causalflow().to_xzcorrections().to_pattern()
-        p_test.infer_pauli_measurements()
-
-        p_ref.remove_pauli_measurements()
-        p_test.remove_pauli_measurements()
+        f_test = p_ref.to_bloch().to_causalflow()
 
         s_ref = p_ref.simulate(rng=rng)
-        s_test = p_test.simulate(rng=rng)
+        s_test = f_test.simulate(rng=rng)
         assert s_ref.isclose(s_test)
 
     # Extract gflow from random circuits
@@ -958,14 +954,10 @@ class TestPattern:
         circuit_1 = rand_circuit(nqubits, depth, rng, use_ccx=False)
         p_ref = circuit_1.transpile().pattern
         p_ref.infer_pauli_measurements()
-        p_test = p_ref.to_bloch().to_gflow().to_xzcorrections().to_pattern()
-        p_test.infer_pauli_measurements()
-
-        p_ref.remove_pauli_measurements()
-        p_test.remove_pauli_measurements()
+        f_test = p_ref.to_bloch().to_gflow()
 
         s_ref = p_ref.simulate(rng=rng)
-        s_test = p_test.simulate(rng=rng)
+        s_test = f_test.simulate(rng=rng)
         assert s_ref.isclose(s_test)
 
     # Extract Pauli flow from random circuits
@@ -978,14 +970,10 @@ class TestPattern:
         circuit_1 = rand_circuit(nqubits, depth, rng, use_ccx=False)
         p_ref = circuit_1.transpile().pattern
         p_ref.infer_pauli_measurements()
-        p_test = p_ref.to_bloch().to_pauliflow().to_xzcorrections().to_pattern()
-        p_test.infer_pauli_measurements()
-
-        p_ref.remove_pauli_measurements()
-        p_test.remove_pauli_measurements()
+        f_test = p_ref.to_bloch().to_pauliflow()
 
         s_ref = p_ref.simulate(rng=rng)
-        s_test = p_test.simulate(rng=rng)
+        s_test = f_test.simulate(rng=rng)
         assert s_ref.isclose(s_test)
 
     @pytest.mark.parametrize("test_case", PATTERN_FLOW_TEST_CASES)
@@ -994,8 +982,8 @@ class TestPattern:
             alpha = 2 * np.pi * fx_rng.random()
             s_ref = test_case.pattern.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
-            p_test = test_case.pattern.to_bloch().to_causalflow().to_xzcorrections().to_pattern()
-            s_test = p_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
+            f_test = test_case.pattern.to_bloch().to_causalflow()
+            s_test = f_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
             assert s_ref.isclose(s_test)
         else:
@@ -1009,8 +997,8 @@ class TestPattern:
             alpha = 2 * np.pi * fx_rng.random()
             s_ref = test_case.pattern.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
-            p_test = test_case.pattern.to_bloch().to_gflow().to_xzcorrections().to_pattern()
-            s_test = p_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
+            f_test = test_case.pattern.to_bloch().to_gflow()
+            s_test = f_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
             assert s_ref.isclose(s_test)
         else:
@@ -1028,8 +1016,8 @@ class TestPattern:
         alpha = 2 * np.pi * fx_rng.random()
         s_ref = test_case.pattern.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
-        p_test = test_case.pattern.to_bloch().to_pauliflow().to_xzcorrections().to_pattern()
-        s_test = p_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
+        f_test = test_case.pattern.to_bloch().to_pauliflow()
+        s_test = f_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
         assert s_ref.isclose(s_test)
 
@@ -1048,11 +1036,11 @@ class TestPattern:
                 4: Measurement.XY(0.4),
             },
         )
-        p_ref = og.to_causalflow().to_xzcorrections().to_pattern()
+        p_ref = og.to_causalflow().to_pattern()
         s_ref = p_ref.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
-        p_test = p_ref.to_causalflow().to_xzcorrections().to_pattern()
-        s_test = p_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
+        f_test = p_ref.to_causalflow()
+        s_test = f_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
         assert s_ref.isclose(s_test)
 
@@ -1072,11 +1060,11 @@ class TestPattern:
             },
         )
 
-        p_ref = og.to_gflow().to_xzcorrections().to_pattern()
+        p_ref = og.to_gflow().to_pattern()
         s_ref = p_ref.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
-        p_test = p_ref.to_gflow().to_xzcorrections().to_pattern()
-        s_test = p_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
+        f_test = p_ref.to_gflow()
+        s_test = f_test.simulate(input_state=PlanarState(Plane.XZ, alpha), rng=fx_rng)
 
         assert s_ref.isclose(s_test)
 
@@ -1091,11 +1079,6 @@ class TestPattern:
         xzc = p_ref.to_xzcorrections()
         xzc.check_well_formed()
         p_test = xzc.to_pattern()
-
-        p_ref.infer_pauli_measurements()
-        p_ref.remove_pauli_measurements()
-        p_test.infer_pauli_measurements()
-        p_test.remove_pauli_measurements()
 
         s_ref = p_ref.simulate(rng=rng)
         s_test = p_test.simulate(rng=rng)
@@ -1203,10 +1186,10 @@ class TestPattern:
         ],
     )
     def test_to_opengraph_roundtrip(self, pattern: Pattern, fx_rng: Generator) -> None:
-        pattern_test = pattern.to_opengraph().to_pattern()
+        og_test = pattern.to_opengraph()
 
         sv = pattern.simulate(rng=fx_rng)
-        sv_test = pattern_test.simulate(rng=fx_rng)
+        sv_test = og_test.simulate(rng=fx_rng)
 
         assert sv.isclose(sv_test)